// features/workflow/types.ts
export interface Node {
  id: string;
  inputs: Record<string, any>;
  class_type: string;
  _meta: Partial<{
    title: string;
    [key: string]: any;
  }> | undefined;
}

export interface Workflow {
  [key: string]: Node;
}

export interface WorkflowRecord {
  id: string;
  name: string;
  serverId: string;
  data: Workflow;
  addMethod: 'file' | 'clipboard' | 'url' | 'preset' | 'server-sync';
  thumbnail?: string;
  createdAt: Date;
  lastUsed?: Date;
  metadata?: Record<string, any>;
}